@@ -21,7 +21,7 @@ import { CacheContext } from './CacheContext';
  * We do a fair bit of pre-processing over them, and these objects hang onto
  * that information.
  */
-export class QueryInfo<TSerialized = GraphSnapshot> {
+export class QueryInfo {
 
   /** The original document (after __typename fields are injected). */
   public readonly document: DocumentNode;
@@ -50,7 +50,7 @@ export class QueryInfo<TSerialized = GraphSnapshot> {
    */
   public readonly variableDefaults: { [Key: string]: JsonValue };
 
-  constructor(context: CacheContext<TSerialized>, raw: RawOperation) {
+  constructor(context: CacheContext, raw: RawOperation) {
     this.document = raw.document;
     this.operation = getOperationOrDie(raw.document);
     this.operationType = this.operation.operation;
