@@ -14,13 +14,16 @@ import { iterParameterized, iterRefs, lazyImmutableDeepSet } from '../util';
  *
  * @throws Will throw an error if there is no corresponding node type
  */
-export function extract<TSerialized>(graphSnapshot: GraphSnapshot, cacheContext: CacheContext<TSerialized>): Serializable.GraphSnapshot {
+export function extract(graphSnapshot: GraphSnapshot, cacheContext: CacheContext): Serializable.GraphSnapshot {
   const result: Serializable.GraphSnapshot = {};
   const entities = graphSnapshot._values;
   // We don't need to check for hasOwnProperty because data._values is
   // created with prototype of 'null'
   for (const id in entities) {
     const nodeSnapshot = entities[id];
+    if (!nodeSnapshot) {
+      continue;
+    }
     const { outbound, inbound, parameterized } = nodeSnapshot;
 
     let type: Serializable.NodeSnapshotType;
