@@ -9,7 +9,7 @@ import { write } from './write';
 /**
  * Return a new graph snapshot pruned to just the shape of the given query
  */
-export function prune<TSerialized>(context: CacheContext<TSerialized>, snapshot: GraphSnapshot, raw: RawOperation) {
+export function prune(context: CacheContext, snapshot: GraphSnapshot, raw: RawOperation) {
   const queryResult = read(context, raw, snapshot, Object.create(null));
   const pruned = write(
     context,
