@@ -1,7 +1,9 @@
 import gql from 'graphql-tag';
-
-import { loadErrorMessageHandler } from '../src/dev/loadErrorMessageHandler';
-import '../src/testing/matchers/index';
+//import "@testing-library/jest-dom";
+import { loadErrorMessageHandler } from '../../apollo-client/src/dev';
+import "../../apollo-client/src/testing/matchers/index";
+import { areApolloErrorsEqual } from "../../apollo-client/src/config/jest/areApolloErrorsEqual";
+import { areGraphQLErrorsEqual } from "../../apollo-client/src/config/jest/areGraphQlErrorsEqual";
 
 // Turn off warnings for repeated fragment names
 gql.disableFragmentWarnings();
@@ -27,5 +29,11 @@ if (!Symbol.asyncDispose) {
     value: Symbol('asyncDispose'),
   });
 }
+
+// @ts-ignore
+expect.addEqualityTesters([areApolloErrorsEqual, areGraphQLErrorsEqual]);
+
+// @ts-ignore
+globalThis.REACT_FALLBACK_THROTTLE_MS = 10;
 
 import './base';
