@@ -1,6 +1,8 @@
 import * as util from 'util';
 
+import { KeyFieldsContext } from '../../apollo-client/src/cache/inmemory/policies';
 import { CacheContext } from '../../src/context';
+import { StoreObject } from '../../apollo-client/src/cache';
 
 export const strictConfig: CacheContext.Configuration = {
   freeze: true,
@@ -13,9 +15,11 @@ export const strictConfig: CacheContext.Configuration = {
     group: jest.fn(),
     groupEnd: jest.fn(),
   },
-  entityIdForNode: (obj) => {
-    const id = obj?.id;
-    return typeof id === 'string' || typeof id === 'number' ? id : undefined;
+  entityIdForNode: (
+    object: Readonly<StoreObject>,
+    context: KeyFieldsContext) => {
+    const id = object?.id;
+    return typeof id === 'string' ? id : typeof id === 'number' ? id.toString() : undefined;
   },
   addTypename: false,
 };
@@ -29,9 +33,11 @@ export const silentConfig: CacheContext.Configuration = {
     group: jest.fn(),
     groupEnd: jest.fn(),
   },
-  entityIdForNode: (obj) => {
-    const id = obj?.id;
-    return typeof id === 'string' || typeof id === 'number' ? id : undefined;
+  entityIdForNode: (
+    object: Readonly<StoreObject>,
+    context: KeyFieldsContext) => {
+    const id = object?.id;
+    return typeof id === 'string' ? id : typeof id === 'number' ? id.toString() : undefined;
   },
   addTypename: false,
 };
