@@ -34,7 +34,7 @@ describe(`writeFragment when using multiple fragments without fragmentName`, () 
           __typename: 'Viewer',
         },
       });
-    }).to.throw(/An error occurred! For more details, see the full error text at /i);
+    }).to.throw(/An error occurred! For more details, see the full error text at |Found 2 fragments. `fragmentName` must be provided when there is not exactly 1 fragment./i);
   });
 
 });
