@@ -47,7 +47,7 @@ describe(`writeFragment with parameterized references`, () => {
       },
       'shipment0': {
         type: Serializable.NodeSnapshotType.EntitySnapshot,
-        inbound: [{ id: [parameterizedId], path: [] }],
+        inbound: [{ id: parameterizedId, path: [] }],
         data: {
           id: 'shipment0',
           destination: 'Seattle',
