@@ -1,6 +1,6 @@
 import { expect } from 'chai';
-import { Cache as CacheInterface } from '@apollo/client';
 
+import { Cache as CacheInterface } from '../../../apollo-client/src';
 import { query, strictConfig } from '../../helpers';
 import { Cache } from '../../../src';
 
